Add new match steper component
