--- pages/matches/new/teams/teams.module.css
@@ -1,3 +1,0 @@
-.footer {
-  text-align: end;
-}
